TF Lite profiling module warning fix profiling module warning fix

+++ tensorflow/lite/kernels/test_util.h
@@ -204,7 +204,7 @@ class SingleOpModel {
     const int channel_index = params->quantized_dimension;
 
     std::vector<int32_t> shape(t->dims->size);
-    for (int i = 0; i < shape.size(); ++i) {
+    for (size_t i = 0; i < shape.size(); ++i) {
       shape[i] = t->dims->data[i];
     }
     const int32_t num_inputs = input_data.size();
